+++ readme.md
@@ -1,1 +1,8 @@
 # Books List
+
+* Simple app where you can add/edit/delete books and import books from Google API. 
+It is possible to filter and sort the books, results are paginated.
+It contains a REST API view, which display a list of books with filtering and sorting options.
+
+###Technology: 
+* Django, Django Rest Framework, Django Crispy Forms, Bootstrap, PostgreSQL, Docker, Docker-Compose,
